Add send functionality
I updated MessageSender class to finish the functionality of sending
automated requests. However, it is not yet tested for large scale send
processes (captchas), nor does it provide good error handling (when an
error occurs, the program schould wait and retry after user input).
In the case of an error, the current HTML page is saved in '/rsc/debug'.

Also, I updated the StealthManager class, by adding a cookie, which is
required for not running into safety prompts before sending requests.