Update error handling in MessageSender
Before the commit, an error would result in a program exit.
Due to the effort of recovering the program and sending the
remaining messages, a better error handling approach was necessary.

After the commit, the program does a prompt and asks
whether to retry sending the message (y\n). Otherwise,
the method returns a false value, which can be counted.